work still incomplete, crashes on some cases of order urgency

+++ Restaurant/Rest/Order.cpp
@@ -144,6 +144,6 @@ void Order::setUrgency(bool urgency) {
 int Order::GetPriority() {
 	return priority;
 }
-void Order::setPriority(float priority) {
+void Order::setPriority(int priority) {
 	this->priority = priority;
 }
